refactor(Controller): 컨트롤러 병합
